requirements.txt
# File: requirements.txt
pandas
sqlalchemy
psycopg2-binary
pymongo
python-dotenv
redis
notebook 
matplotlib 
seaborn
flask
Flask-SocketIO